+++ bin/universal-webpack.js
@@ -15,13 +15,6 @@ if (!command_line_arguments.settings)
 
 var settings = require(path.resolve(process.cwd(), command_line_arguments.settings))
 
-var webpack_configuration
-
-if (command_line_arguments.config)
-{
-	webpack_configuration = require(path.resolve(process.cwd(), command_line_arguments.config))
-}
-
 if (command_line_arguments._.length === 0)
 {
 	return usage('No command specified')
@@ -37,7 +30,7 @@ var command = command_line_arguments._[0]
 switch (command)
 {
 	case 'prepare':
-		universal_webpack.prepare(settings, webpack_configuration)
+		universal_webpack.prepare(settings, command_line_arguments.basePath)
 		break
 
 	default:
@@ -66,11 +59,12 @@ function usage(reason)
 	console.log('             in that "--watch"ed folder and therefore won\'t')
 	console.log('             restart the server on code changes.')
 	console.log('             ')
-	console.log('             An optional --config flag may be specified')
-	console.log('             in which case `context` will be read from this webpack config')
-	console.log('             and will be used to resolve the paths in')
-	console.log('             `./universal-webpack-settings.js` which are resolved')
-	console.log('             from `process.cwd()` by default.')
+	console.log('             An optional --basePath argument may be passed')
+	console.log('             (same as Webpack\'s `.context` setting)')
+	console.log('             which is gonna be the base path for resolving')
+	console.log('             the `server` path inside `./universal-webpack-settings.js`.')
+	console.log('             Is `process.cwd()` by default, i.e. the folder')
+	console.log('             from which the `universal-webpack` command is being run.')
 
 	if (reason)
 	{