+++ test/loaders.js
@@ -1,11 +1,20 @@
 import chai from 'chai'
 chai.should()
 
-import { find_style_rules, loader_name_filter, parse_loader, stringify_loader, normalize_rule_loaders } from '../source/loaders'
+import
+{
+	get_style_rules,
+	find_loader,
+	parse_loader,
+	stringify_loader,
+	normalize_rule_loaders,
+	normalize_configuration_rule_loaders
+}
+from '../source/loaders'
 
 describe(`webpack loader utilities`, function()
 {
-	it(`should find style loaders`, function()
+	it(`should get style rules`, function()
 	{
 		let configuration
 
@@ -42,7 +51,7 @@ describe(`webpack loader utilities`, function()
 			}
 		}
 
-		find_style_rules(configuration).should.deep.equal
+		get_style_rules(configuration).should.deep.equal
 		([{
 			use:
 			[{
@@ -51,14 +60,92 @@ describe(`webpack loader utilities`, function()
 		}])
 	})
 
-	it(`should detect style loader`, function()
+	it(`should get style rules when using "oneOf"`, function()
 	{
-		loader_name_filter('style')('style-loader').should.equal(true)
-		loader_name_filter('style')('style-loader?query=true&gay=porn').should.equal(true)
-		loader_name_filter('style')('style').should.equal(true)
-		loader_name_filter('style')('style?query=true').should.equal(true)
+		let configuration
 
-		loader_name_filter('style')('style_loader').should.equal(false)
+		configuration =
+		{
+			module:
+			{
+				rules:
+				[{
+					test: /\.css$/,
+					oneOf:
+					[{
+						use:
+						[{
+							loader: 'not-a-style-loader'
+						}]
+					},
+					{
+						use:
+						[{
+							loader: 'not-a-style-loader'
+						},
+						{
+							loader: 'style-loader'
+						}]
+					}]
+				},
+				{
+					test: /\.css$/,
+					oneOf:
+					[{
+						use:
+						[{
+							loader: 'style-loader'
+						}]
+					},
+					{
+						use:
+						[{
+							loader: 'not-a-style-loader'
+						}]
+					}]
+				}]
+			}
+		}
+
+		get_style_rules(configuration).should.deep.equal
+		([{
+			use:
+			[{
+				loader: 'not-a-style-loader'
+			},
+			{
+				loader: 'style-loader'
+			}]
+		},
+		{
+			use:
+			[{
+				loader: 'style-loader'
+			}]
+		}])
+	})
+
+	it(`should find style loader`, function()
+	{
+		find_loader
+		({
+			test: /\.css$/,
+			use:
+			[{
+				loader: 'after-loader'
+			},
+			{
+				loader: 'style-loader'
+			},
+			{
+				loader: 'before-loader'
+			}]
+		},
+		'style-loader')
+		.should.deep.equal
+		({
+			loader: 'style-loader'
+		})
 	})
 
 	it(`should parse loaders`, function()
@@ -158,6 +245,44 @@ describe(`webpack loader utilities`, function()
 		let execute = () => normalize_rule_loaders(loader)
 		execute.should.throw(`You have both a compound ".loader" and a ".query"`)
 
+		// Should recurse into `oneOf`
+
+		loader =
+		{
+			test: /.css$/,
+			oneOf:
+			[{
+				resourceQuery: /hot/, // foo.css?hot
+				use: 'hot-style-loader'
+			},
+			{
+				resourceQuery: /gay/, // foo.css?gay
+				use: 'gay-style-loader'
+			}]
+		}
+
+		normalize_rule_loaders(loader)
+
+		loader.should.deep.equal
+		({
+			test: /.css$/,
+			oneOf:
+			[{
+				resourceQuery: /hot/, // foo.css?hot
+				use:
+				[{
+					loader: 'hot-style-loader'
+				}]
+			},
+			{
+				resourceQuery: /gay/, // foo.css?gay
+				use:
+				[{
+					loader: 'gay-style-loader'
+				}]
+			}]
+		})
+
 		// No `loader` is specified
 
 		loader =
@@ -170,7 +295,7 @@ describe(`webpack loader utilities`, function()
 		}
 
 		execute = () => normalize_rule_loaders(loader)
-		execute.should.throw(`Neither "loaders" nor "loader" nor "use" are present inside a module rule`)
+		execute.should.throw(`Neither "loaders" nor "loader" nor "use" nor "oneOf" are present inside a module rule`)
 
 		// Convert compound `loader` to `use` array
 
@@ -199,6 +324,38 @@ describe(`webpack loader utilities`, function()
 					a: 'b'
 				}
 			}]
+		})
+	})
+
+	it('should normalize configuration rule loaders', function()
+	{
+		const configuration =
+		{
+			module:
+			{
+				rules:
+				[{
+					test: /.css$/,
+					use: 'css-loader'
+				}]
+			}
+		}
+
+		normalize_configuration_rule_loaders(configuration)
+
+		configuration.should.deep.equal
+		({
+			module:
+			{
+				rules:
+				[{
+					test: /.css$/,
+					use:
+					[{
+						loader: 'css-loader'
+					}]
+				}]
+			}
 		})
 	})
 })