+++ test/server configuration.js
@@ -3,11 +3,19 @@ chai.should()
 
 import path from 'path'
 import webpack from 'webpack'
-import server, { replace_style_loader, dont_emit_file_loader, extract_package_name } from '../source/server configuration'
+
+import server,
+{
+	replace_style_loader,
+	dont_emit_file_loader,
+	extract_package_name,
+	is_external
+}
+from '../source/server configuration'
 
 describe(`server configuration`, function()
 {
-	it(`should add "emit: false" for "file-loader" and "url-loader"`, function()
+	it(`should add "emitFile: false" for "file-loader" and "url-loader"`, function()
 	{
 		const configuration =
 		{
@@ -28,6 +36,23 @@ describe(`server configuration`, function()
 						loader: 'url-loader',
 						options: { limit: 10000 }
 					}]
+				},
+				{
+					test: /\.png$/,
+					oneOf:
+					[{
+						use:
+						[{
+							loader: 'url-loader',
+							options: { limit: 10000 }
+						}]
+					},
+					{
+						use:
+						[{
+							loader: 'whatever-loader'
+						}]
+					}]
 				}]
 			}
 		}
@@ -45,6 +70,12 @@ describe(`server configuration`, function()
 			loader: 'url-loader',
 			options: { limit: 10000, emitFile: false }
 		}])
+
+		configuration.module.rules[2].oneOf[0].use.should.deep.equal
+		([{
+			loader: 'url-loader',
+			options: { limit: 10000, emitFile: false }
+		}])
 	})
 
 	it(`should replace "style-loader" and "css-loader" with "css-loader/locals" on the server side`, function()
@@ -54,14 +85,47 @@ describe(`server configuration`, function()
 			module:
 			{
 				rules:
+				// Generic
 				[{
 					test: /\.css$/,
 					use:
 					[{
+						loader: 'after-loader'
+					},
+					{
 						loader: 'style-loader'
 					},
 					{
 						loader: 'css-loader'
+					},
+					{
+						loader: 'before-loader'
+					}]
+				},
+				// "oneOf"
+				{
+					test: /\.css$/,
+					oneOf:
+					[{
+						use:
+						[{
+							loader: 'after-loader'
+						},
+						{
+							loader: 'style-loader'
+						},
+						{
+							loader: 'css-loader'
+						},
+						{
+							loader: 'before-loader'
+						}]
+					},
+					{
+						use:
+						[{
+							loader: 'whatever-loader'
+						}]
 					}]
 				}]
 			}
@@ -69,9 +133,28 @@ describe(`server configuration`, function()
 
 		replace_style_loader(configuration)
 
+		// Generic
 		configuration.module.rules[0].use.should.deep.equal
 		([{
+			loader: 'after-loader'
+		},
+		{
 			loader: 'css-loader/locals'
+		},
+		{
+			loader: 'before-loader'
+		}])
+
+		// "oneOf"
+		configuration.module.rules[1].oneOf[0].use.should.deep.equal
+		([{
+			loader: 'after-loader'
+		},
+		{
+			loader: 'css-loader/locals'
+		},
+		{
+			loader: 'before-loader'
 		}])
 	})
 
@@ -102,5 +185,103 @@ describe(`server configuration`, function()
 		extract_package_name('@incomplete').should.equal('@incomplete')
 		extract_package_name('@private/dependency').should.equal('@private/dependency')
 		extract_package_name('@private/dependency/lib/anything').should.equal('@private/dependency')
+	})
+
+	it(`should check if a dependency is external`, function()
+	{
+		// Generic NPM package
+		is_external
+		(
+			'react-responsive-ui/modules/Select',
+			{},
+			{}
+		)
+		.should.equal(true)
+
+		// Starts with `!`
+		is_external
+		(
+			'!react-responsive-ui/modules/Select',
+			{},
+			{}
+		)
+		.should.equal(false)
+
+		// Starts with `-!`
+		is_external
+		(
+			'-!react-responsive-ui/modules/Select',
+			{},
+			{}
+		)
+		.should.equal(false)
+
+		// Invalid NPM package name
+		is_external
+		(
+			'react:responsive;ui/modules/Select',
+			{},
+			{}
+		)
+		.should.equal(false)
+
+		// `resolve.alias`
+		is_external
+		(
+			'react-responsive-ui/modules/Select',
+			{
+				resolve:
+				{
+					alias:
+					{
+						'react-responsive-ui': '/react-responsive-ui'
+					}
+				}
+			},
+			{}
+		)
+		.should.equal(false)
+
+		// Is a loaded asset
+		is_external
+		(
+			'react-responsive-ui/modules/Select.css',
+			{},
+			{}
+		)
+		.should.equal(false)
+
+		// Not a loaded asset
+		is_external
+		(
+			'react-responsive-ui/modules/Select.css',
+			{},
+			{
+				loadExternalModuleFileExtensions: []
+			}
+		)
+		.should.equal(true)
+
+		// `excludeFromExternals` string
+		is_external
+		(
+			'react-responsive-ui/modules/Select',
+			{},
+			{
+				excludeFromExternals: ['react-responsive-ui']
+			}
+		)
+		.should.equal(false)
+
+		// `excludeFromExternals` regular expression
+		is_external
+		(
+			'react-responsive-ui/modules/Select',
+			{},
+			{
+				excludeFromExternals: [/^react-responsive-ui(\/.*)?$/]
+			}
+		)
+		.should.equal(false)
 	})
 })